+++ lib/td_bg/cache/domain_loader.ex
@@ -19,22 +19,22 @@ defmodule TdBg.Cache.DomainLoader do
     GenServer.call(__MODULE__, {:refresh, domain_id})
   end
 
+  def refresh_deleted do
+    GenServer.cast(__MODULE__, :refresh_deleted)
+  end
+
   def delete(domain_id) do
     GenServer.call(__MODULE__, {:delete, domain_id})
   end
 
   @impl true
   def init(state) do
-    unless Application.get_env(:td_bg, :env) == :test do
-      schedule_work(:load_cache, 0)
-    end
-
     {:ok, state}
   end
 
   @impl true
   def handle_call({:refresh, :all}, _from, state) do
-    load_all_domains()
+    load_domains()
     {:reply, :ok, state}
   end
 
@@ -51,14 +51,9 @@ defmodule TdBg.Cache.DomainLoader do
   end
 
   @impl true
-  def handle_info(:load_cache, state) do
-    load_all_domains()
-
+  def handle_cast(:refresh_deleted, state) do
+    remove_deleted_domains()
     {:noreply, state}
-  end
-
-  defp schedule_work(action, millis) do
-    Process.send_after(self(), action, millis)
   end
 
   defp load_domain(domain_id) do
@@ -68,9 +63,30 @@ defmodule TdBg.Cache.DomainLoader do
     |> load_domain_data()
   end
 
-  defp load_all_domains do
+  defp load_domains do
     Taxonomies.list_domains()
     |> load_domain_data()
+  end
+
+  defp remove_deleted_domains do
+    {results, errors} =
+      Taxonomies.list_domains(deleted: true)
+      |> Enum.map(fn %{id: id} -> TaxonomyCache.delete_domain(id) end)
+      |> Enum.split_with(fn {:ok, _} -> true end)
+
+    if length(errors) > 0 do
+      Logger.warn("Cache deletion failed with #{length(errors)} errors")
+    else
+      remove_count =
+        Enum.reduce(results, 0, fn
+          {:ok, [_, _, _, _, _, n]}, acc when is_integer(n) -> acc + n
+          _, acc -> acc
+        end)
+
+      if remove_count > 0 do
+        Logger.info("Removed #{remove_count} deleted domains")
+      end
+    end
   end
 
   def load_domain_data(domains) do